feat: 휴양지 Top 5 Suspense 범위 축소, 스켈레톤 UI 추가
